requirements.txt
cffi==1.14.5
cryptography==3.4.7
pycparser==2.20
pytest==6.2.4
pyyaml==5.4.1
